fix(correctness): float truncation in weights, quorum denominator guard bypass
MemberRepository::sumNotDeletedVoteWeight() returned int, silently
truncating fractional voting_power (e.g. 1.5 → 1). Now returns float,
matching sumActiveWeight(). DashboardController present_weight cast
also fixed (int → float).

VoteEngine/QuorumEngine/QuorumController: max(1, $eligibleMembers)
prevented the existing $denominator <= 0 guard from firing when zero
members are eligible. Removed the max(1, …) so the guard correctly
sets quorumMet = false instead of computing a bogus ratio against
denominator 1.

+++ app/Repository/MemberRepository.php
@@ -113,11 +113,11 @@ class MemberRepository extends AbstractRepository {
     /**
      * Sum of voting_power for non-deleted members (for dashboard eligible weight).
      */
-    public function sumNotDeletedVoteWeight(string $tenantId): int {
-        return (int) ($this->scalar(
+    public function sumNotDeletedVoteWeight(string $tenantId): float {
+        return (float) ($this->scalar(
             'SELECT COALESCE(SUM(COALESCE(voting_power, 1.0)), 0) FROM members WHERE tenant_id = :tid AND deleted_at IS NULL',
             [':tid' => $tenantId],
-        ) ?? 0);
+        ) ?? 0.0);
     }
 
     /**
